[miq_provision_virt_workflow.rb] Avoid respond_to?
respond_to? calls in mass are slow since they require some object
introspection, where doing a simple `nil` check is much quicker.

By including just the `cloud_tenant_id` in the SELECT clause, it allows
for the swapping of `respond_to?` with an `nil` check, with the
functionality that existed preserved.

Benchmark
---------

**Before**

|     ms | queries | query (ms) |   rows |
|   ---: |    ---: |       ---: |   ---: |
|  28440 |      33 |     1929.5 | 243133 |
|  28491 |      33 |     1933.6 | 243133 |
|  28858 |      33 |     2631.9 | 243133 |

**After**

|     ms | queries | query (ms) |   rows |
|   ---: |    ---: |       ---: |   ---: |
|  27107 |      33 |     1958.5 | 243133 |
|  26803 |      33 |     1944.2 | 243133 |
|  27642 |      33 |     1965.5 | 243133 |