Remove the worker restart_interval
We already have memory limits to prevent workers from growing out of
control so the pre-emptive restart_interval is at best overkill.

The restart interval also causes issues with queueing full refreshes on
refresh_worker startup.  To prevent too many full refreshes we don't run
a full on VMware RefreshWorker start, but this causes other issues in
that when a second EMS is added it doesn't actually do a full refresh
and you have to do it manually.

Also all of our limit checks are supported by systemd except this one.