[GitRepository] Add #git_transaction method
And also defines the following private helper methods and constants:

- `LOCKFILE_DIR`         # Where the locks are stored
- `#git_lock_filename`   # name of the lockfile for this record
- `#acquire_git_lock`    # fetches a lock for the instance (one per)
- `#release_git_lock`    # releases existing lock for other processes

`#git_transaction` is meant to be a multi-purpose method for dealing
with any file system changes so that only one process is able to
interact with it at a time.  According to the docs (`ri File#flock`),
`File#flock(FILE::LOCK_EX)` should block until the lock is available for
the process, so the first process to acquire the lock will process
first.

The method is also designed so that nested `#git_transaction` calls will
no-op and not release the lock pre-maturely.  This is important for
running multiple actions that by design need to happen by a single
process, but the sub actions are also configured to use the lock.

Implementation of this method will be done in a followup commit, but the
method is also meant to be a public interface so that when existing
methods implement this method, locks will work just fine with user
defined `git_transaction` blocks as well as with GitRepository method
API.