[EmbeddedAnsible] Filter out non-playbooks
When syncing playbooks from the GitRepository, make sure we are only
pulling in files that match a playbook extension (yaml), and have the
"looks of a playbook".  Basically, they should include one of the
following lines:

  "- hosts: ..."
  "- import_playbook: ..."
  "- include:"

(sans-quotes)

So if there is a `hosts` line, or we are importing another playbook or
role, it is a playbook.

Logic for this commit borrowed from AWX:

  https://github.com/ansible/awx/blob/128fa894/awx/main/utils/ansible.py#L17
  https://github.com/ansible/awx/blob/128fa894/awx/main/utils/ansible.py#L39-L66

Partially fixes https://bugzilla.redhat.com/show_bug.cgi?id=1734446